Significantly speed up the fancy xdoc viewer. (fixes issue 134)
Previously, the jump-to box was rendering every :short string (XML -> TEXT) at
xdoc startup time.  This (embarassingly simple) patch avoids this, and does the
rendering lazily, only as needed, so the page loads way faster.  As an added
bonus, we no longer need the hogan library.

This is a viewer-only change that should be compatible with previously
generated manuals, e.g., for ACL2 6.4.

I suspect it isn't worth the time to tinker with the format of xindex.js for
now (which is sort of a shame, because I spent a fair bit of time thinking up
how to do that).  We can eventually tweak it if we need to, but I think this
patch will make performance acceptable for most users.  (Of course, please
report any experiences to the contrary...)