Define: updates to :returns specifiers
1. Return spec theorems now often have more natural names, i.e.,

     (define foo (x)
       :returns (ans natp ...)
       ...)

Will now create a theorem named NATP-OF-FOO instead of RETURN-TYPE-OF-FOO.
For multiple-value returns the situation is similar, e.g.,

     (define foo (x)
       :returns (mv (a natp ...)
                    (b consp ...))
       ...)

Will now produce, NATP-OF-FOO.A and CONSP-OF-FOO.B, instead of
RETURN-TYPE-OF-FOO.A and RETURN-TYPE-OF-FOO.B.

For complex return specifiers, we will still generate these "generic"
return-type names.

2. Return spec theorems can now be explicitly named.  For instance:

   (define foo (x)
       :returns (new-x (equal (len new-x) (len x))
                       :name len-of-foo)
       ...)

3. A new macro, MORE-RETURNS, allows you to submit additional return
specifiers.  For instance

  (define foo ((x ...))
    :returns (base natp :rule-classes :type-prescription)
    ...
    ///
    (more-returns
      (base print-base-p :hyp :guard)
      (base (<= 2 base) :rule-classes :linear
            :name foo-lower-bound)))

You can also use this outside of the /// area, by naming foo.  See
the documentation for details.

Perhaps this gives us enough control over when to prove return specs.

Update issue 184