std/define: unlocalize calls of set-define-current-function (fixes issue 190).
Thanks for the bug report.

It looks like the problem is that we are only locally (in a table) setting the
name of the function being defined.  So, if the book isn't certified, that
local event isn't available and it isn't getting its name.

Making the table events non-local only slows down vl/parsetree.lisp from
1m12.829s to 1m13.996 seconds, which could well be noise.  So I think it's fine
to just unlocalize these.