Added (declaim (inline min-fixnum)) as suggested by Jared.
Fixed comment in split-at-position (thanks to Jared for pointing out a
mistake in that comment).

Fixed a comment typo.

Updated doc.lisp to match books/system/doc/acl2-doc.lisp.

Improved a warning in chk-acceptable-linear-rule2 to clarify that only
relevant triggers are a problem for free variables.

Added function find-symbol-in-current-package, and used it to speed up (we
think) some printing routines.  Thanks to Jared Davis for pointing out
potential value of such a change.