Documentation experiment.
This is a fairly aggressive attmept to reorganize many ACL2 topics, mostly from
acl2-built-ins, into related groups such as alists, arrays, characters, errors,
numbers, packages, etc.

This reorganization is done in doc/top.lisp:fix-the-hierarchy, so it should
only affect the web-based manual, and it should be easy to undo if people do
not think it is progress.

This is very much a half-assed first draft, and there are a lot of things that
obviously need a lot of work.  For instance, it's probably not sensible to have
all of these things directly underneath ACL2.  I'd appreciate feedback on the
general idea and how to improve upon it.

Also: hack to improve strings/pretty-program certification time.